Move common classes to util package
