+++ src/main/java/possibletriangle/skygrid/compat/jei/ingredient/DimensionIngredient.java
@@ -2,7 +2,6 @@ package possibletriangle.skygrid.compat.jei.ingredient;
 
 import mezz.jei.api.ingredients.IIngredientType;
 import net.minecraft.world.dimension.DimensionType;
-import possibletriangle.skygrid.data.loading.DimensionConfig;
 
 public enum DimensionIngredient implements IIngredientType<DimensionType> {
