+++ src/main/java/possibletriangle/skygrid/util/provider/OffsetBlock.java
@@ -1,4 +1,4 @@
-package possibletriangle.skygrid.provider;
+package possibletriangle.skygrid.util.provider;
 
 import net.minecraft.util.math.BlockPos;
