+++ src/main/java/possibletriangle/skygrid/util/provider/property/PropertyProvider.java
@@ -1,4 +1,4 @@
-package possibletriangle.skygrid.provider.property;
+package possibletriangle.skygrid.util.provider.property;
 
 import net.minecraft.block.BlockState;
 import net.minecraft.state.IProperty;
