+++ src/main/java/possibletriangle/skygrid/util/provider/property/SetProperty.java
@@ -1,4 +1,4 @@
-package possibletriangle.skygrid.provider.property;
+package possibletriangle.skygrid.util.provider.property;
 
 import net.minecraft.state.IProperty;
