+++ src/main/java/possibletriangle/skygrid/world/custom/CustomBiomeProvider.java
@@ -6,6 +6,7 @@ import net.minecraft.world.biome.provider.BiomeProvider;
 import net.minecraft.world.gen.OverworldGenSettings;
 import net.minecraft.world.gen.layer.Layer;
 import net.minecraft.world.gen.layer.LayerUtil;
+import possibletriangle.skygrid.util.loading.CreateOptions;
 
 import java.util.function.Supplier;
