+++ src/main/java/possibletriangle/skygrid/world/custom/CustomChunkGenerator.java
@@ -7,6 +7,7 @@ import net.minecraft.world.gen.ChunkGenerator;
 import net.minecraft.world.gen.GenerationSettings;
 import net.minecraft.world.gen.Heightmap;
 import net.minecraft.world.gen.WorldGenRegion;
+import possibletriangle.skygrid.util.loading.CreateOptions;
 
 import java.util.function.Supplier;
